+++ source/gatery/export/vhdl/CodeFormatting.h
@@ -42,11 +42,11 @@ class CodeFormatting
 		inline const std::string &getIndentation() const { return m_indentation; }
 		inline const std::string &getFileHeader() const { return m_fileHeader; }
 		inline const std::string &getFilenameExtension() const { return m_filenameExtension; }
-		void indent(std::ostream &stream, unsigned depth) const;
+		void indent(std::ostream &stream, size_t depth) const;
 		virtual void formatEntityComment(std::ostream &stream, const std::string &entityName, const std::string &comment) = 0;
 		virtual void formatBlockComment(std::ostream &stream, const std::string &blockName, const std::string &comment) = 0;
-		virtual void formatProcessComment(std::ostream &stream, unsigned indentation, const std::string &processName, const std::string &comment) = 0;
-		virtual void formatCodeComment(std::ostream &stream, unsigned indentation, const std::string &comment) = 0;
+		virtual void formatProcessComment(std::ostream &stream, size_t indentation, const std::string &processName, const std::string &comment) = 0;
+		virtual void formatCodeComment(std::ostream &stream, size_t indentation, const std::string &comment) = 0;
 
 		virtual void formatConnectionType(std::ostream &stream, const VHDLSignalDeclaration &declaration) = 0;
 		virtual void formatDeclaration(std::ostream &stream, const VHDLSignalDeclaration &declaration) = 0;
@@ -65,17 +65,17 @@ class CodeFormatting
 			SIG_CONSTANT,
 		};
 
-		virtual std::string getNodeName(const hlim::BaseNode *node, unsigned attempt) const = 0;
-		virtual std::string getSignalName(const std::string &desiredName, SignalType type, unsigned attempt) const = 0;
-		virtual std::string getPackageName(const std::string &desiredName, unsigned attempt) const = 0;
-		virtual std::string getEntityName(const std::string &desiredName, unsigned attempt) const = 0;
-		virtual std::string getBlockName(const std::string &desiredName, unsigned attempt) const = 0;
-		virtual std::string getProcessName(const std::string &desiredName, bool clocked, unsigned attempt) const = 0;
-		virtual std::string getClockName(const std::string &desiredName, unsigned attempt) const = 0;
-		virtual std::string getIoPinName(const std::string &desiredName, unsigned attempt) const = 0;
-		virtual std::string getInstanceName(const std::string &desiredName, unsigned attempt) const = 0;
+		virtual std::string getNodeName(const hlim::BaseNode *node, size_t attempt) const = 0;
+		virtual std::string getSignalName(const std::string &desiredName, SignalType type, size_t attempt) const = 0;
+		virtual std::string getPackageName(const std::string &desiredName, size_t attempt) const = 0;
+		virtual std::string getEntityName(const std::string &desiredName, size_t attempt) const = 0;
+		virtual std::string getBlockName(const std::string &desiredName, size_t attempt) const = 0;
+		virtual std::string getProcessName(const std::string &desiredName, bool clocked, size_t attempt) const = 0;
+		virtual std::string getClockName(const std::string &desiredName, size_t attempt) const = 0;
+		virtual std::string getIoPinName(const std::string &desiredName, size_t attempt) const = 0;
+		virtual std::string getInstanceName(const std::string &desiredName, size_t attempt) const = 0;
 /*
-		virtual void instantiateExternal(std::ostream &stream, const hlim::Node_External *node, unsigned indent,
+		virtual void instantiateExternal(std::ostream &stream, const hlim::Node_External *node, size_t indent,
 										 const std::vector<std::string> &inputSignalNames, const std::vector<std::string> &outputSignalNames, const std::vector<std::string> &clockNames) const = 0;
 */
 	protected:
@@ -89,31 +89,31 @@ class CodeFormatting
 class DefaultCodeFormatting : public CodeFormatting
 {
 	public:
-		//using ExternalNodeHandler = std::function<bool(const CodeFormatting*, std::ostream &, const hlim::Node_External *, unsigned, const std::vector<std::string> &, const std::vector<std::string> &, const std::vector<std::string> &)>;
+		//using ExternalNodeHandler = std::function<bool(const CodeFormatting*, std::ostream &, const hlim::Node_External *, size_t, const std::vector<std::string> &, const std::vector<std::string> &, const std::vector<std::string> &)>;
 
 		DefaultCodeFormatting();
 
-		virtual std::string getNodeName(const hlim::BaseNode *node, unsigned attempt) const override;
-		virtual std::string getSignalName(const std::string &desiredName, SignalType type, unsigned attempt) const override;
-		virtual std::string getPackageName(const std::string &desiredName, unsigned attempt) const override;
-		virtual std::string getEntityName(const std::string &desiredName, unsigned attempt) const override;
-		virtual std::string getBlockName(const std::string &desiredName, unsigned attempt) const override;
-		virtual std::string getProcessName(const std::string &desiredName, bool clocked, unsigned attempt) const override;
-		virtual std::string getClockName(const std::string &desiredName, unsigned attempt) const override;
-		virtual std::string getIoPinName(const std::string &desiredName, unsigned attempt) const override;
-		virtual std::string getInstanceName(const std::string &desiredName, unsigned attempt) const override;
+		virtual std::string getNodeName(const hlim::BaseNode *node, size_t attempt) const override;
+		virtual std::string getSignalName(const std::string &desiredName, SignalType type, size_t attempt) const override;
+		virtual std::string getPackageName(const std::string &desiredName, size_t attempt) const override;
+		virtual std::string getEntityName(const std::string &desiredName, size_t attempt) const override;
+		virtual std::string getBlockName(const std::string &desiredName, size_t attempt) const override;
+		virtual std::string getProcessName(const std::string &desiredName, bool clocked, size_t attempt) const override;
+		virtual std::string getClockName(const std::string &desiredName, size_t attempt) const override;
+		virtual std::string getIoPinName(const std::string &desiredName, size_t attempt) const override;
+		virtual std::string getInstanceName(const std::string &desiredName, size_t attempt) const override;
 
 		virtual void formatEntityComment(std::ostream &stream, const std::string &entityName, const std::string &comment) override;
 		virtual void formatBlockComment(std::ostream &stream, const std::string &blockName, const std::string &comment) override;
-		virtual void formatProcessComment(std::ostream &stream, unsigned indentation, const std::string &processName, const std::string &comment) override;
-		virtual void formatCodeComment(std::ostream &stream, unsigned indentation, const std::string &comment) override;
+		virtual void formatProcessComment(std::ostream &stream, size_t indentation, const std::string &processName, const std::string &comment) override;
+		virtual void formatCodeComment(std::ostream &stream, size_t indentation, const std::string &comment) override;
 
 		virtual void formatConnectionType(std::ostream &stream, const VHDLSignalDeclaration &declaration) override;
 		virtual void formatDeclaration(std::ostream &stream, const VHDLSignalDeclaration &declaration) override;
 		virtual void formatDataType(std::ostream &stream, VHDLDataType dataType) override;
 /*
 		void addExternalNodeHandler(ExternalNodeHandler nodeHandler);
-		virtual void instantiateExternal(std::ostream &stream, const hlim::Node_External *node, unsigned indent,
+		virtual void instantiateExternal(std::ostream &stream, const hlim::Node_External *node, size_t indent,
 										 const std::vector<std::string> &inputSignalNames, const std::vector<std::string> &outputSignalNames, const std::vector<std::string> &clockNames) const override;
 */
 	protected:
